feat(tty): docs

+++ src/pty.rs
@@ -41,7 +41,7 @@ impl PtyMaster {
                     if libc::setsid() == -1 {
                         return Err(std::io::Error::last_os_error());
                     }
-                    if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY, 0) == -1 {
+                    if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY as libc::c_ulong, 0) == -1 {
                         return Err(std::io::Error::last_os_error());
                     }
                     Ok(())
@@ -77,7 +77,13 @@ pub fn set_window_size(fd: impl AsFd, rows: u16, cols: u16) -> std::io::Result<(
     };
     // Safety: fd is a valid file descriptor (enforced by AsFd),
     // ws is a valid winsize struct on the stack.
-    let ret = unsafe { libc::ioctl(fd.as_fd().as_raw_fd(), libc::TIOCSWINSZ, &ws) };
+    let ret = unsafe {
+        libc::ioctl(
+            fd.as_fd().as_raw_fd(),
+            libc::TIOCSWINSZ as libc::c_ulong,
+            &ws,
+        )
+    };
     if ret == -1 {
         Err(std::io::Error::last_os_error())
     } else {
